python/requirements.txt
uvicorn~=0.22.0
fastapi==0.98.0
python-dotenv~=1.0.0
langchain==0.0.232
pydantic~=1.10.7
requests==2.31.0
PyYAML~=6.0
pytest~=7.3.1
pytest-mock~=3.11.1
pytest-dotenv~=0.5.2
setuptools==67.8.0
openai~=0.27.8
mysqlclient~=2.1.1
psycopg2-binary~=2.9.6
pytest-manual-marker~=2.0.0.0
cohere~=4.11.2
beautifulsoup4~=4.12.2
chromadb~=0.3.26
tiktoken~=0.4.0
llama-index~=0.6.35
weaviate-client~=3.21.0
astunparse~=1.6.3
httpx~=0.24.1
